+++ modules/processing/geotk-processing-math/src/main/java/org/geotoolkit/process/math/min/MinProcess.java
@@ -37,9 +37,9 @@ public class MinProcess extends AbstractProcess {
 
         final Double[] set = value(SET, inputParameters);
 
-        Double min = Math.min(set[0], set[1]);
+        double min = Math.min(set[0].doubleValue(), set[1].doubleValue());
         for (int i=1; i<set.length; i++) {
-            min = Math.min(min.doubleValue(), set[i].doubleValue());
+            min = Math.min(min, set[i].doubleValue());
         }
 
         getOrCreate(RESULT_NUMBER, outputParameters).setValue(min);
